Add support for onlyTrashed model scope. Fix #928 Move withTrashed method to Eloquent engine.
