+++ src/Engines/BaseEngine.php
@@ -104,14 +104,6 @@ abstract class BaseEngine implements DataTableEngineContract
      * @var bool
      */
     protected $autoFilter = true;
-    
-    /**
-     * Select trashed records in count function for models with soft deletes trait.
-     * By default we do not select soft deleted records
-     *
-     * @var bool
-     */
-    protected $withTrashed = false;
 
     /**
      * Callback to override global search.
@@ -401,19 +393,6 @@ abstract class BaseEngine implements DataTableEngineContract
     {
         $this->columnDef['escape'] = $columns;
 
-        return $this;
-    }
-    
-    /**
-     * Change withTrashed flag value.
-     *
-     * @param bool $withTrashed
-     * @return $this
-     */
-    public function withTrashed($withTrashed = true)
-    {
-        $this->withTrashed = $withTrashed;
-        
         return $this;
     }
